+++ src/main/java/org/moodle/api/client/moodle_rest_client/infrastructure/outputs/dtos/course/course/CourseDateDTO.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class CourseDateDTO {
     private String label;
-    private Integer timestamp;
-    private Integer relativeto;
+    private Long timestamp;
+    private Long relativeto;
     private String dataid;
 }
